structural changes for plotting tasks
