+++ columnflow/tasks/framework/mixins.py
@@ -781,52 +781,6 @@ class PlotMixin(AnalysisTask):
         description="a command to execute after the task has run to visualize plots right in the "
         "terminal; no default",
     )
-    skip_ratio = luigi.BoolParameter(
-        default=False,
-        significant=False,
-        description="when True, no ratio (usually Data/Bkg ratio) is drawn in the lower panel; "
-        "default: False",
-    )
-    shape_norm = luigi.BoolParameter(
-        default=False,
-        significant=False,
-        description="when True, each process is normalized on it's integral in the upper panel; "
-        "default: False",
-    )
-    skip_legend = luigi.BoolParameter(
-        default=False,
-        significant=False,
-        description="when True, no legend is drawn; default: False",
-    )
-    skip_cms = luigi.BoolParameter(
-        default=False,
-        significant=False,
-        description="when True, no CMS logo is drawn; default: False",
-    )
-    y_scale = luigi.ChoiceParameter(
-        choices=(law.NO_STR, "linear", "log"),
-        default=law.NO_STR,
-        significant=False,
-        description="string parameter to define the y-axis scale of the plot in the upper panel; "
-        "choices: NO_STR,linear,log; no default",
-    )
-    process_lines = law.CSVParameter(
-        default=(),
-        description="comma-separated process names or patterns to plot as individual lines; "
-        "can also be the key of a mapping defined in the 'process_groups' auxiliary data of "
-        "the config; uses no process when empty; empty default",
-    )
-
-    def get_plot_parameters(self):
-        # convert parameters to usable values during plotting
-        return {
-            "skip_ratio": self.skip_ratio,
-            "shape_norm": self.shape_norm,
-            "skip_legend": self.skip_legend,
-            "skip_cms": self.skip_cms,
-            "yscale": None if self.y_scale == law.NO_STR else self.y_scale,
-            "process_lines": self.process_lines,
-        }
 
     def get_plot_func(self, func_name: str) -> Callable:
         """
