+++ columnflow/tasks/framework/plotting.py~
@@ -1,0 +1,6 @@
+# coding: utf-8
+
+"""
+Base tasks for different types of plotting tasks
+"""
+
